Finish checkout process for k-stuff

+++ komanda/app/models/product.rb
@@ -2,7 +2,7 @@ class Product
   include Mongoid::Document
   
   field :name,  type: String
-  field :price, type: Float
+  field :price, type: Integer
   field :pictures, type: Array, default: []
   field :num_sold, default: 0
